refactor: Move role_holder_addresses to ietf.api.views
